+++ xmake.lua
@@ -3,7 +3,7 @@ add_rules("mode.debug", "mode.release")
 add_repositories("levimc-repo https://github.com/LiteLDev/xmake-repo.git")
 
 if is_config("target_type", "server") then
-    add_requires("levilamina 26.40.2", {configs = {target_type = "server"}})
+    add_requires("levilamina 26.51.1", {configs = {target_type = "server"}})
 else
     add_requires("levilamina 26.40.2", {configs = {target_type = "client"}})
 end